# Vectorize CharacterMap.IndexOfOpeningCharacter