@MihaZupan 2022-04-21
Some good news, this now turned into a discussion about where we can improve the existing helper methods in .NET: https://github.com/dotnet/runtime/issues/68328